Update CMake configuration, modify Fibonacci function for memoization, and add build script

+++ src/include/fibApp.hpp
@@ -1,4 +1,4 @@
 namespace fibApp
 {
-    int fibApp(int num);
+    float fibApp(float num);
 } // namespace fibApp
